backend/requirements.txt
# Core FastAPI
fastapi
uvicorn[standard]

# Supabase client
supabase

# Configurações
pydantic-settings
python-dotenv

# Utilitários HTTP
httpx
python-multipart
